adding mail page in setting
