+++ tailwind.config.cjs
@@ -34,7 +34,8 @@ const config = {
           "neutral-content": "#ffffff",
           "base-100": "#1A1A1B",
           "base-200": "#2D2D2E",
-          "base-300": "#424243",
+          "base-300": "#59595A",
+          "base-400": "#59595A",
           "base-content": "#ebecf0",
           info: "#66c6ff",
           success: "#87d039",
